server/requirements.txt
-i https://pypi.org/simple
aniso8601==7.0.0
asgiref==3.2.7
django-filter==2.2.0
django-mysql==3.3.0
django==3.0.4
djangorestframework==3.11.0
graphene-django==2.9.0
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
gunicorn==20.0.4
mysqlclient==1.4.6
promise==2.3
pytz==2019.3
rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
